refactor(CardSlot): move out from PVZ.h
