+++ pvzclass/pvzclass.vcxproj.filters
@@ -836,5 +836,8 @@
     <ClInclude Include="Classes\Challenge.hpp">
       <Filter>头文件\Classes</Filter>
     </ClInclude>
+    <ClInclude Include="Classes\SeedBank.hpp">
+      <Filter>头文件\Classes</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
